move SoilType to shared domain
