+++ apps/web/src/features/create-project/views/soils/soils-surface-areas/index.tsx
@@ -1,7 +1,3 @@
-import {
-  RECOMMENDED_M2_PER_KWC_FOR_ACCESS_PATHS,
-  RECOMMENDED_M2_PER_KWC_FOR_FOUNDATIONS,
-} from "../../domain/photovoltaic";
 import SoilDistributionForm from "./SoilsSurfaceAreasForm";
 
 import {
@@ -9,6 +5,10 @@ import {
   ProjectCreationStep,
   setSoilsSurfaceAreas,
 } from "@/features/create-project/application/createProject.reducer";
+import {
+  RECOMMENDED_M2_PER_KWC_FOR_ACCESS_PATHS,
+  RECOMMENDED_M2_PER_KWC_FOR_FOUNDATIONS,
+} from "@/features/create-project/domain/photovoltaic";
 import {
   useAppDispatch,
   useAppSelector,
