+++ apps/web/src/features/create-site/application/siteSoilsCarbonStorage.reducer.spec.ts
@@ -1,7 +1,7 @@
-import { SoilType } from "../domain/siteFoncier.types";
 import { SoilsCarbonStorageMock } from "../infrastructure/soils-carbon-storage-service/soilsCarbonStorageMock";
 import { fetchCarbonStorageForSoils } from "./siteSoilsCarbonStorage.actions";
 
+import { SoilType } from "@/shared/domain/soils";
 import { createStore } from "@/store";
 import { getTestAppDependencies } from "@/test/testAppDependencies";
