+++ apps/web/src/features/create-site/domain/generateSiteDesignation.spec.ts
@@ -1,5 +1,7 @@
 import { generateSiteDesignation } from "./generateSiteDesignation";
-import { SiteDraft, SoilType } from "./siteFoncier.types";
+import { SiteDraft } from "./siteFoncier.types";
+
+import { SoilType } from "@/shared/domain/soils";
 
 const buildSiteDraft = (siteData: Partial<SiteDraft>): SiteDraft => {
   return {
