+++ apps/web/src/features/create-site/views/soils/soils-surface-areas/SoilsSurfaceAreasForm.tsx
@@ -2,8 +2,8 @@ import { useMemo } from "react";
 import { useForm } from "react-hook-form";
 import Button from "@codegouvfr/react-dsfr/Button";
 import { Input } from "@codegouvfr/react-dsfr/Input";
-import { SoilType } from "../../../domain/siteFoncier.types";
 
+import { SoilType } from "@/shared/domain/soils";
 import { formatNumberFr } from "@/shared/services/format-number/formatNumber";
 import { getLabelForSoilType } from "@/shared/services/label-mapping/soilTypeLabelMapping";
 import SliderNumericInput from "@/shared/views/components/form/NumericInput/SliderNumericInput";
