+++ apps/web/src/shared/domain/soils.ts
@@ -1,0 +1,19 @@
+export enum SoilType {
+  BUILDINGS = "BUILDINGS",
+  IMPERMEABLE_SOILS = "IMPERMEABLE_SOILS",
+  MINERAL_SOIL = "MINERAL_SOIL",
+  ARTIFICIAL_GRASS_OR_BUSHES_FILLED = "ARTIFICIAL_GRASS_OR_BUSHES_FILLED",
+  ARTIFICIAL_TREE_FILLED = "ARTIFICIAL_TREE_FILLED",
+  FOREST_DECIDUOUS = "FOREST_DECIDUOUS",
+  FOREST_CONIFER = "FOREST_CONIFER",
+  FOREST_POPLAR = "FOREST_POPLAR",
+  FOREST_MIXED = "FOREST_MIXED",
+  PRAIRIE_GRASS = "PRAIRIE_GRASS",
+  PRAIRIE_BUSHES = "PRAIRIE_BUSHES",
+  PRAIRIE_TREES = "PRAIRIE_TREES",
+  ORCHARD = "ORCHARD", // verger
+  CULTIVATION = "CULTIVATION", // culture
+  VINEYARD = "VINEYARD", // vigne
+  WET_LAND = "WET_LAND", // zone humide
+  WATER = "WATER", // plan d'eau
+}
