+++ apps/web/src/shared/services/label-mapping/soilTypeLabelMapping.ts
@@ -1,4 +1,4 @@
-import { SoilType } from "@/features/create-site/domain/siteFoncier.types";
+import { SoilType } from "@/shared/domain/soils";
 
 export const getLabelForSoilType = (value: SoilType): string => {
   switch (value) {
